[Platinum V] Title: 공장, Time: 304 ms, Memory: 15700 KB -BaekjoonHub

+++ 백준/Platinum/7578. 공장/공장.cc
@@ -43,7 +43,7 @@ int main()
 	//문제해결
 	ll ans = 0;
 	for (int i = 1; i <= n; ++i) {
-		ans += (ll)Query(1, 1, n, B[i], n);
+		ans += Query(1, 1, n, B[i], n);
 		Connect(1, 1, n, B[i]);
 	}
